requirements.txt
llama-index
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
langchain
langchain-openai
langgraph
streamlit
python-dotenv